requirements.txt
acme==1.1.0
amqp==2.6.0
appdirs==1.4.4
apturl==0.5.2
arabic-reshaper==2.1.1
asgiref==3.2.7
astroid==2.3.3
atomicwrites==1.4.0
attrs==19.3.0
backports.entry-points-selectable==1.1.0
bcrypt==3.1.7
billiard==3.6.3.0
blinker==1.4
Brlapi==0.7.0
ccsm==0.9.14.1
celery==4.4.4
certbot==0.40.0
certbot-apache==0.39.0
certifi==2020.4.5.1
cffi==1.14.4
chardet==3.0.4
chrome-gnome-shell==0.0.0
Click==7.0
colorama==0.4.3
command-not-found==0.3
compizconfig-python==0.9.14.1
ConfigArgParse==0.13.0
configobj==5.0.6
coreapi==2.3.3
coreschema==0.0.4
coverage==5.1
cryptography==2.8
cupshelpers==1.0
cx-Freeze==6.1
cycler==0.10.0
dbus-python==1.2.16
decorator==4.4.2
decouple==0.0.7
defer==1.0.6
defusedxml==0.6.0
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
dj-database-url==0.5.0
Django==3.0.7
django-allauth==0.41.0
django-cors-headers==3.3.0
django-debug-toolbar==3.2.2
django-extensions==2.2.9
django-filter==21.1
django-formtools==2.3
django-guardian==2.2.0
django-phone-field==1.8.0
django-post-office==3.4.1
django-rest-auth==0.9.5
django-utils-six==2.0
djangorestframework==3.11.0
djangorestframework-api-key==2.0.0
djangorestframework-bulk==0.2.1
djangorestframework-filters==0.11.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
docopt==0.6.2
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
future==0.18.2
gevent==20.9.0
greenlet==0.4.17
Guake==3.6.3
gunicorn==20.0.4
html5lib==1.1
httpie==2.1.0
httplib2==0.14.0
idna==2.9
import-ipynb==0.1.3
importlib-metadata==1.6.1
isort==4.3.4
itypes==1.2.0
Jinja2==2.11.2
josepy==1.2.0
jsonfield==3.1.0
keyring==18.0.1
kiwisolver==1.0.1
kombu==4.6.10
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
lockfile==0.12.2
logilab-common==1.4.3
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.1.2
mccabe==0.6.1
mininet==2.3.0
mininet-wifi==2.6
mock==3.0.5
monotonic==1.5
more-itertools==8.2.0
mysqlclient==2.0.3
netifaces==0.10.4
nftables==0.1
num2words==0.5.10
numpy==1.17.4
oauthlib==3.1.0
olefile==0.46
ovs==2.13.3
packaging==20.3
paramiko==2.6.0
parsedatetime==2.4
pbr==5.4.5
pexpect==4.6.0
Pillow==7.1.2
platformdirs==2.3.0
pluggy==0.13.1
protobuf==3.6.1
psutil==5.5.1
psycopg2==2.8.5
psycopg2-binary==2.8.5
py==1.8.1
pycairo==1.16.2
pycodestyle==2.5.0
pycparser==2.20
pycups==1.9.73
pyflakes==2.1.1
Pygments==2.6.1
PyGObject==3.36.0
PyICU==2.4.2
PyJWT==1.7.1
pylint==2.4.4
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.7
PyPDF2==1.26.0
pyRFC3339==1.1
pytest==5.4.1
pytest-cov==2.8.1
python-apt==2.0.0+ubuntu0.20.4.6
python-augeas==0.5.0
python-bidi==0.4.2
python-dateutil==2.8.1
python-debian===0.1.36ubuntu1
python-decouple==3.3
python3-openid==3.1.0
pytz==2020.1
pyxdg==0.26
PyYAML==5.3.1
redis==3.5.3
reportlab==3.5.56
requests==2.23.0
requests-oauthlib==1.3.0
requests-toolbelt==0.8.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
selinux==3.0
simplejson==3.16.0
six==1.15.0
slip==0.6.5
slip.dbus==0.6.5
somepackage==1.2.3
sortedcontainers==2.1.0
sqlparse==0.3.1
ssh-import-id==5.10
systemd-python==234
twilio==6.49.0
ubuntu-advantage-tools==27.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==1.25.9
usb-creator==0.3.7
vboxapi==1.0
vine==1.3.0
virtualenv==20.0.23
virtualenvwrapper-win==1.2.6
wadllib==1.3.3
wcwidth==0.1.9
webencodings==0.5.1
whitenoise==5.0.1
wrapt==1.11.2
xhtml2pdf==0.2.5
xkit==0.0.0
zipp==3.1.0
zope.component==4.3.0
zope.event==4.5.0
zope.hookable==5.0.0
zope.interface==5.2.0
